refactored FuzzTests into a separate project
