+++ fuzz-tests/Tests/ReadOnlySpan/ReadOnlySpanTests.Linq.cs
@@ -1,0 +1,9 @@
+using static SpanExtensions.FuzzTests.TestHelper;
+
+namespace SpanExtensions.FuzzTests
+{
+    public static class ReadOnlySpanLinqTests
+    {
+
+    }
+}
